[Week9][Improvement] Refactor project structure
