+++ Week9/app/src/main/java/com/example/marvelcharacters/app/Injection.kt
@@ -1,12 +1,15 @@
 package com.example.marvelcharacters.app
 
-import com.example.marvelcharacters.networking.RemoteApi
-import com.example.marvelcharacters.repository.CharacterDao
-import com.example.marvelcharacters.repository.CharacterRepository
-import com.example.marvelcharacters.repository.RoomRepository
+import androidx.work.WorkManager
+import com.example.marvelcharacters.repository.remote.RemoteApi
+import com.example.marvelcharacters.repository.local.CharacterDao
+import com.example.marvelcharacters.repository.local.CharacterRepository
+import com.example.marvelcharacters.repository.local.RoomRepository
 
 object Injection {
-    fun provideRepository(): CharacterRepository = RoomRepository()
+    fun provideRepository(): CharacterRepository =
+        RoomRepository()
     fun provideRemoteApi(): RemoteApi = App.remoteApi
     fun provideCharacterDao(): CharacterDao = App.characterDb.characterDao()
+    fun provideWorkManager(): WorkManager = App.workManager
 }
