+++ Week9/app/src/main/java/com/example/marvelcharacters/model/entities/Character.kt
@@ -2,7 +2,7 @@ package com.example.marvelcharacters.model.entities
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import com.example.marvelcharacters.repository.DATABASE_NAME
+import com.example.marvelcharacters.repository.local.DATABASE_NAME
 import kotlinx.serialization.Serializable
 
 @Serializable
