+++ Week9/app/src/main/java/com/example/marvelcharacters/model/entities/Thumbnail.kt
@@ -1,6 +1,5 @@
 package com.example.marvelcharacters.model.entities
 
-import androidx.room.TypeConverter
 import kotlinx.serialization.Serializable
 
 @Serializable
