+++ Week9/app/src/main/java/com/example/marvelcharacters/repository/local/CharacterDatabase.kt
@@ -1,7 +1,6 @@
-package com.example.marvelcharacters.repository
+package com.example.marvelcharacters.repository.local
 
 import androidx.room.Database
-import androidx.room.Room
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
 import com.example.marvelcharacters.model.converters.Converters
