+++ Week9/app/src/main/java/com/example/marvelcharacters/repository/local/CharacterRepository.kt
@@ -1,4 +1,4 @@
-package com.example.marvelcharacters.repository
+package com.example.marvelcharacters.repository.local
 
 import androidx.lifecycle.LiveData
 import com.example.marvelcharacters.model.entities.Character
