+++ Week9/app/src/main/java/com/example/marvelcharacters/repository/remote/NetworkStatusChecker.kt
@@ -1,4 +1,4 @@
-package com.example.marvelcharacters.networking
+package com.example.marvelcharacters.repository.remote
 
 import android.net.ConnectivityManager
 import android.net.NetworkCapabilities
