+++ Week9/app/src/main/java/com/example/marvelcharacters/repository/remote/RemoteApi.kt
@@ -1,12 +1,11 @@
-package com.example.marvelcharacters.networking
+package com.example.marvelcharacters.repository.remote
 
 import com.example.marvelcharacters.model.entities.Character
-import com.example.marvelcharacters.model.response.*
 
 
 const val BASE_URL = "https://gateway.marvel.com"
 
-class RemoteApi(private val apiService: RemoteApiService) {
+class RemoteApi(private val apiService: RemoteDataSource) {
 
     suspend fun getCharacters(): Result<List<Character>> = try {
         val response = apiService.getCharacters()
